+++ src/Callbacks/CreateSelectOptionValueCallback.php
@@ -77,11 +77,11 @@ class CreateSelectOptionValueCallback extends AbstractCallback
     /**
      * Will be invoked by a observer it has been registered for.
      *
-     * @param \TechDivision\Import\Observers\ObserverInterface $observer The observer
+     * @param \TechDivision\Import\Observers\AttributeCodeAndValueAwareObserverInterface|null $observer The observer
      *
      * @return mixed The modified value
      */
-    public function handle(AttributeCodeAndValueAwareObserverInterface $observer)
+    public function handle(AttributeCodeAndValueAwareObserverInterface $observer = null)
     {
 
         // set the observer