+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ## Features
 
 * Remove deprecated classes and methods
+* Add techdivision/import#162
 * Add techdivision/import-cli-simple#216
 * Add techdivision/import-configuration-jms#25
 * Remove unnecessary identifiers from configuration