+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# Version 17.0.1
+
+## Bugfixes
+
+* Fixed techdivision/import-attribute#47
+
+## Features
+
+* None
+
 # Version 17.0.0
 
 ## Bugfixes